BACKEND/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
cloudinary==1.38.0
django-cloudinary-storage==0.3.0
Pillow==10.2.0
django-ratelimit==4.1.0
python-decouple==3.8
